+++ src/pulumi/core-shared/security/config_registry.py
@@ -143,17 +143,21 @@ for assignment in key_vault_config.get("iam", {}).get("role_assignments", []):
     user_group_ref_key = assignment.get("user_group_ref_key")
     if user_group_ref_key is not None:
         GroupRoleAssignment(
+            principal_name=user_group_ref_key,
+            principal_id=user_groups[user_group_ref_key]["object_id"],
             role_name=assignment["role_definition_name"],
-            group_object_id=user_groups[user_group_ref_key]["object_id"],
             scope=key_vault.id,
+            scope_description="config-registry",
         )
 
 # Grant access to the Automation Service Principal to manage the key vault.
 # We are going to be creating secrets in the key vault later on, so we need the access.
 ServicePrincipalRoleAssignment(
-    service_principal_object_id=azure_client.object_id,
+    principal_id=azure_client.object_id,
+    principal_name="automation-service-principal",
     role_name="Key Vault Administrator",
     scope=key_vault.id,
+    scope_description="config-registry",
 )
 
 # ----------------------------------------------------------------------------------------------------------------------